Create calendar subpackage and CalendarServiceBase
